requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
configparser==5.0.0
Distance==0.1.3
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
minio==5.0.8
pdf2image==1.10.0
Pillow==6.2.1
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2019.3
reportlab==3.5.32
requests==2.23.0
six==1.14.0
uuid==1.30
urllib3==1.25.8
Werkzeug==1.0.1